+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 [Unreleased]: https://github.com/sounisi5011/package-version-git-tag/compare/v3.0.0...master
 
 * Drop support for Node.js 10 and 12
+* Fix edge cases in CLI argument parsing process
+* Update dev dependencies
 
 ### Supported Node version
 
@@ -26,6 +28,10 @@ Because it is the minimum version available for Vitest.
 * [#200] - Migrate from AVA to Vitest
 
 [contributor:johnschult]: https://github.com/johnschult
+
+### Fixed
+
+* [#210] - Fix edge cases in CLI argument parsing process
 
 ### Documentation
 
@@ -104,6 +110,7 @@ Because it is the minimum version available for Vitest.
 
     This is needed for vitest.
 * [#200] - `vitest@0.29.2`
+* [#210] - `vitest-mock-process@1.0.4`
 
 ### Removed Dependencies
 
@@ -160,6 +167,7 @@ Because it is the minimum version available for Vitest.
 [#202]: https://github.com/sounisi5011/package-version-git-tag/pull/202
 [#207]: https://github.com/sounisi5011/package-version-git-tag/pull/207
 [#208]: https://github.com/sounisi5011/package-version-git-tag/pull/208
+[#210]: https://github.com/sounisi5011/package-version-git-tag/pull/210
 
 ## [3.0.0] (2020-06-02 UTC)